refactor(editor): project-based routing /project/[id]/
- / → Project list (sample + empty, New disabled until DB)
- /project/[id]/diagram → Diagram editor
- /project/[id]/specs, bom, connections → management pages
- loadProject(id): reset + load by ID (sample or empty)
- NavBar links project-relative
- Remove old routes, fix monorepo lint
