+++ apps/editor/src/routes/+layout.svelte
@@ -1,17 +1,12 @@
 <script lang="ts">
   import { Tooltip } from 'bits-ui'
-  import NavBar from '$lib/components/NavBar.svelte'
-  import { diagramState, initDarkMode } from '$lib/context.svelte'
+  import { initDarkMode } from '$lib/context.svelte'
   import '../app.css'
 
   let { children } = $props()
 
   $effect(() => {
     return initDarkMode()
-  })
-
-  $effect(() => {
-    diagramState.initialize()
   })
 </script>
 
@@ -20,7 +15,4 @@
   <meta name="description" content="Interactive network topology diagram editor">
 </svelte:head>
 
-<Tooltip.Provider delayDuration={200}>
-  <NavBar />
-  {@render children()}
-</Tooltip.Provider>
+<Tooltip.Provider delayDuration={200}> {@render children()} </Tooltip.Provider>
