+++ apps/editor/src/routes/project/[id]/(content)/+layout.svelte
